+++ ruby-client-sdk/lib/sdk/requestbodies.rb
@@ -2749,6 +2749,45 @@ module OpenApiSDK
       res
     end
 
+    sig { params(request: Operations::RequestBodyPutMultipartDifferentFileNameRequestBody).returns(Utils::FieldAugmented) }
+    def request_body_put_multipart_different_file_name(request)
+
+      url, params = @sdk_configuration.get_server_details
+      base_url = Utils.template_url(url, params)
+      url = "#{base_url}/anything/requestBodies/put/multipart/differentFileName"
+      headers = {}
+      req_content_type, data, form = Utils.serialize_request_body(request, :request, :multipart)
+      headers['content-type'] = req_content_type
+      raise StandardError, 'request body is required' if data.nil? && form.nil?
+      headers['Accept'] = 'application/json'
+      headers['x-speakeasy-user-agent'] = "speakeasy-sdk/#{@sdk_configuration.language} #{@sdk_configuration.sdk_version} #{@sdk_configuration.gen_version} #{@sdk_configuration.openapi_doc_version}"
+
+      r = @sdk_configuration.client.put(url) do |req|
+        req.headers = headers
+        Utils.configure_request_security(req, @sdk_configuration.security) if !@sdk_configuration.nil? && !@sdk_configuration.security.nil?
+        if form
+          req.body = Utils.encode_form(form)
+        elsif Utils.match_content_type(req_content_type, 'application/x-www-form-urlencoded')
+          req.body = URI.encode_www_form(data)
+        else
+          req.body = data
+        end
+      end
+
+      content_type = r.headers.fetch('Content-Type', 'application/octet-stream')
+
+      res = Operations::RequestBodyPutMultipartDifferentFileNameResponse.new(
+        status_code: r.status, content_type: content_type, raw_response: r
+      )
+      if r.status == 200
+        if Utils.match_content_type(content_type, 'application/json')
+          out = Utils.unmarshal_complex(r.env.response_body, Operations::RequestBodyPutMultipartDifferentFileNameRes)
+          res.res = out
+        end
+      end
+      res
+    end
+
     sig { params(request: Operations::RequestBodyPutMultipartFileRequestBody).returns(Utils::FieldAugmented) }
     def request_body_put_multipart_file(request)