+++ ruby-client-sdk/lib/sdk/sdkconfiguration.rb
@@ -53,7 +53,7 @@ module OpenApiSDK
       @globals = globals.nil? ? {} : globals
       @language = 'ruby'
       @openapi_doc_version = '0.1.0'
-      @sdk_version = '1.33.3'
+      @sdk_version = '1.33.4'
       @gen_version = '2.131.1'
     end