+++ RELEASES.md
@@ -736,3 +736,18 @@ Based on:
 - [typescript v1.35.3] typescript-client-sdk
 ### Releases
 - [Go v1.35.3] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.35.3 - go-client-sdk
+
+## 2023-09-28 20:10:58
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 
+- Speakeasy CLI 1.91.2 (2.131.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [php v1.33.4] php-client-sdk
+- [python v1.34.4] python-client-sdk
+- [ruby v1.33.4] ruby-client-sdk
+- [typescript v1.35.4] typescript-client-sdk
+- [go v1.35.4] go-client-sdk
+- [java v1.33.4] java-client-sdk
+### Releases
+- [Go v1.35.4] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.35.4 - go-client-sdk